Revert "Test possible sheet behaviour regression bugfix"
This reverts commit 6e2734cf5f68aaeb40c9e8ffc4efadd1a431edf0.

`const actorData = context.actor.data;` was used only briefly in
the Boilerplate system and was reverted to the original
`const actorData = this.actor.data.toObject(false);` for the v9 update.
QWFVTT used the `context...` version as that was briefly current.
It seems that both work, but the `this...` variant is (more) canonical.

See https://gitlab.com/asacolips-projects/foundry-mods/boilerplate/
commit history in:

- 8fbac1af9f99ddafb01a9760deb10981b96bce2c
- 82352ee737df0896b633549baefd65ea908daa7f
- f622c5791bb419d7411bd6bc05aed0ce8132c8da